+++ CT_ROOT/system/modules/hofff_calendarfield/config/config.php
@@ -1,0 +1,6 @@
+<?php
+
+/**
+ * Form fields
+ */
+$GLOBALS['TL_FFL']['calendarfield'] = 'Hofff\Contao\Calendarfield\FormCalendarField';