+++ CT_ROOT/system/modules/hofff_calendarfield/languages/de/default.php
@@ -1,0 +1,14 @@
+<?php
+
+/**
+ * Misc
+ */
+$GLOBALS['TL_LANG']['MSC']['calendarfield_tooltip'] = 'Datum auswählen';
+
+/**
+ * Errors
+ */
+$GLOBALS['TL_LANG']['ERR']['calendarfield_direction_ltToday'] = 'Bitte geben Sie ein Datum in der Vergangenheit (exkl. heute) ein.';
+$GLOBALS['TL_LANG']['ERR']['calendarfield_direction_leToday'] = 'Bitte geben Sie ein Datum in der Vergangenheit (inkl. heute) ein.';
+$GLOBALS['TL_LANG']['ERR']['calendarfield_direction_geToday'] = 'Bitte geben Sie ein Datum in der Zukunft (inkl. heute) ein.';
+$GLOBALS['TL_LANG']['ERR']['calendarfield_direction_gtToday'] = 'Bitte geben Sie ein Datum in der Zukunft (exkl. heute) ein.';