+++ contao/config/config.php
@@ -1,6 +1,0 @@
-<?php
-
-/**
- * Form fields
- */
-$GLOBALS['TL_FFL']['jcalendar'] = 'Hofff\Contao\JCalendarfield\FormJCalendarField';